backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
defusedxml==0.7.1
django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.9.0
pyopenssl==25.1.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
twisted==25.5.0
txaio==23.1.1
typing-extensions==4.14.0
urllib3==2.4.0
zope-interface==7.2
